src/requirements.txt
paho-mqtt
yeelight
pyyaml
pycrypto
